% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{assess takes catch and index data, initial values for the parameters,
and a flag saying whether to compute uncertainty estimates for the model parameters,
fits the model to the data, and returns fitted parameter estimates, biomass estimates,
the nll, and any convergence diagnostics}
\usage{
assess(catch, index, calc.vcov = FALSE, pars.init)
}
\arguments{
\item{catch}{vector of catch data}

\item{index}{vector of index data}

\item{pars.init}{initial values for the parameters}

\item{calc.vcoc}{flag computing uncertainty estimates}
}
\description{
assess takes catch and index data, initial values for the parameters,
and a flag saying whether to compute uncertainty estimates for the model parameters,
fits the model to the data, and returns fitted parameter estimates, biomass estimates,
the nll, and any convergence diagnostics
}
