% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control}
\alias{control}
\title{control takes the estimated biomass and control rule parameters and returns
harvest takes the current biomass, a catch, and the model parameters to compute next year's biomass}
\usage{
control(estimated.biomass, control.pars)
}
\arguments{
\item{estimated.biomass}{estimated biomass from assess function}

\item{control.pars}{control rule parameters}
}
\description{
control takes the estimated biomass and control rule parameters and returns
harvest takes the current biomass, a catch, and the model parameters to compute next year's biomass
}
