% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{dynamics}
\alias{dynamics}
\title{dynamics takes the model parameters, a time series of catches,}
\usage{
dynamics(pars, C, yrs)
}
\arguments{
\item{pars}{model parameters}

\item{C}{time series of catch}

\item{yrs}{years to compute biomass over}
}
\description{
dynamics takes the model parameters, a time series of catches,
}
