% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate takes the parameters & historical biomass trajectories of the operating model,
the specifications of the harvest control rule,}
\usage{
evaluate(
  pars.iter,
  biomass.iter,
  control.pars,
  data.years,
  proj.years,
  iterations,
  ...
)
}
\arguments{
\item{pars.iter}{operating model parameters from operating model}

\item{biomass.iter}{historical biomass trajectories from operating model}

\item{control.pars}{parameters defining the harvest control rule}

\item{data.years}{years that have data informing the operating model}

\item{proj.years}{years to do projections over}

\item{iterations}{number of iterations}
}
\description{
evaluate takes the parameters & historical biomass trajectories of the operating model,
the specifications of the harvest control rule,
}
