% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement.R
\name{implement}
\alias{implement}
\title{implement takes the TAC and returns an adjusted TAC based on implementation error
(over- or under-fishing)}
\usage{
implement(TAC, ...)
}
\arguments{
\item{TAC}{total allowable catch defined by control function (harvest control rule)}
}
\description{
implement takes the TAC and returns an adjusted TAC based on implementation error
(over- or under-fishing)
}
