% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{nll}
\alias{nll}
\title{nll takes the parameters, catches, and index data and returns the nll of the fit
to the dynamics function}
\usage{
nll(pars, C, U)
}
\arguments{
\item{pars}{parameter values for the dynamics model}

\item{C}{catch data}

\item{U}{index data}
}
\description{
nll takes the parameters, catches, and index data and returns the nll of the fit
to the dynamics function
}
