% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{observe}
\alias{observe}
\title{observe takes the true biomass from the operating model and adds observation error,
defined by sigma}
\usage{
observe(biomass, sigma)
}
\arguments{
\item{biomass}{biomass from the operating model in the function dynamics}

\item{sigma}{observation error}
}
\description{
observe takes the true biomass from the operating model and adds observation error,
defined by sigma
}
