% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionplot.R
\name{projectionplot}
\alias{projectionplot}
\title{projectionplot takes the projection results and plots them}
\usage{
projectionplot(project.results)
}
\arguments{
\item{projection.results}{biomass projections into the future}
}
\description{
projectionplot takes the projection results and plots them
}
