% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\name{schaefer}
\alias{schaefer}
\title{schaefer takes the current biomass, a catch, and the model parameters to compute next year's biomass}
\usage{
schaefer(B, C, K, r)
}
\arguments{
\item{B}{define}

\item{C}{define}

\item{K}{define}

\item{r}{define}
}
\description{
schaefer takes the current biomass, a catch, and the model parameters to compute next year's biomass
}
