% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\name{schaefer}
\alias{schaefer}
\title{schaefer takes the current biomass, a catch, and the model parameters to compute next year's biomass}
\usage{
schaefer(B, C, K, r)
}
\arguments{
\item{B}{current year biomass}

\item{C}{current year catch}

\item{K}{carrying capacity}

\item{r}{population growth rate}
}
\description{
schaefer takes the current biomass, a catch, and the model parameters to compute next year's biomass
}
